AI_Body_Size_Measurement_Project/ai_model/requirements.txt
# AI Model Requirements
# Separate from backend requirements for AI-specific dependencies

# Core ML/AI
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
pillow==10.1.0

# Scientific Computing
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Image Processing
albumentations==1.3.1
imageio==2.33.0

# Web Framework for Model Serving
flask==3.0.0
flask-cors==4.0.0

# Model Optimization
onnxruntime==1.16.3
onnx==1.15.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# GPU Support (Optional - uncomment if using GPU)
# tensorflow-gpu==2.15.0
# torch-cuda==2.1.1
